Patch for Julia 0.5.1.
This commit allows the quickstart code to run on the latest version of
Julia (0.5.1). Quite a few methods and some syntax has been deprecated
in the last two years. This patch replaces these with the equivalent
new methods and syntax.

Included changes:
- Rewrite deprecated Dict declarations.
- Use String instead of deprecated ASCIIString.
- Use Int instead of deprecated int.
- Use ceil(Integer, x) instead of deprecated iceil(x).
- Use floor(Integer, x) instead of deprecated ifloor(x).
- Import PyPlot to automatically resolve matplotlib dependency.
- Explicitly convert E, tua_s and tua_p to Integer when needed. (nasty!)
- Use Range instead of Range1. (typo?)

Tested using http://cyrusmaher.github.io/CauseMap.jl/Quickstart.html. I
did not test with any other data or settings.

The graph output looks aesthetically different, but seems okay in terms
of the data. I suspect this is the result of changes to matplotlib, but
I did not investigate further.